+++ src/Category_Colors/CSS/Pro.php
@@ -131,9 +131,20 @@ class Pro {
 	}
 
 	/**
+	 * Make event spacer transparent.
+	 *
+	 * @return void
+	 */
+	public static function fix_pro_spacer_background() {
+		$css[] = 'tribe-events-pro-week-grid__multiday-event-wrapper tribe-events-pro-week-grid__multiday-event--empty';
+		$css[] = '{ background-color: transparent !important; }';
+		self::echo_css( $css );
+	}
+
+	/**
 	 * Fix double left border on all day week view event.
 	 *
-	 * @param string $slug
+	 * @param string $slug Category slug.
 	 *
 	 * @return void
 	 */