+++ CHANGES.md
@@ -1,6 +1,7 @@
 #### [unreleased]
 * rename title to correspond to The Events Calendar standard
 * retructure CSS additions to not end with comma, added in `category.css.php`
+* fix empty spacer classes to be transparent
 
 #### 7.0.6 / 2022-05-20
 * move all Pro CSS to `CSS\Pro`